requirements.txt
numpy>=1.24.0
onnx>=1.14.0
onnxruntime>=1.15.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
